CRAFTING SYSTEM & MORE
WHOOOOOOOO BOY we got a lot going on:

Player now has an inventory and a hotbar which can hold craftable items in the game. player can select items to move them around the inventory

Map generator can easily add more types of block to be generated in the map. Ore is placed around the map with miners that use the drunkards walk algorithm.

Camera follow player around the screen.

Added a crafting system that will populate with recipes so long as the resources are in the  players ore sack.

Player can place items around the world via the world grid. Player can destory items, which drops itemdrops that the player can pick up and it is inserted back into their inventory